@@ -52,32 +52,36 @@
         <project.build.sourceEncoding>${utf8}</project.build.sourceEncoding>
         <project.reporting.outputEncoding>${utf8}</project.reporting.outputEncoding>
         <java.version>1.8</java.version>
-        <spring-boot.version>2.2.5.RELEASE</spring-boot.version>
-        <spring-boot-admin.version>2.2.2</spring-boot-admin.version>
-        <mybatis-plus-boot-starter.version>3.3.1</mybatis-plus-boot-starter.version>
+
+        <spring-boot.version>2.4.3</spring-boot.version>
+        <spring-boot-admin.version>2.3.1</spring-boot-admin.version>
+        <mybatis-plus-boot-starter.version>3.4.2</mybatis-plus-boot-starter.version>
+        <mybatis-plus-generator.version>3.4.1</mybatis-plus-generator.version>
         <velocity.version>2.2</velocity.version>
+
         <commons-lang3.version>3.9</commons-lang3.version>
         <commons-collections4.version>4.4</commons-collections4.version>
         <commons-io.version>2.6</commons-io.version>
-        <commons-codec.version>1.14</commons-codec.version>
-        <commons-net.version>3.6</commons-net.version>
-        <commons-pool2.version>2.8.0</commons-pool2.version>
-        <commons-text.version>1.8</commons-text.version>
-        <fastjson.version>1.2.67</fastjson.version>
-        <mysql.version>5.1.47</mysql.version>
-        <reflections.version>0.9.9</reflections.version>
+        <commons-codec.version>1.15</commons-codec.version>
+        <commons-net.version>3.7.2</commons-net.version>
+        <commons-pool2.version>2.9.0</commons-pool2.version>
+        <commons-text.version>1.9</commons-text.version>
+
+        <fastjson.version>1.2.75</fastjson.version>
+        <mysql.version>5.1.49</mysql.version>
+        <reflections.version>0.9.12</reflections.version>
         <jansi.version>1.18</jansi.version>
-        <lombok.version>1.18.12</lombok.version>
-        <hutool.version>5.2.4</hutool.version>
-        <junit.version>4.12</junit.version>
+
+        <lombok.version>1.18.18</lombok.version>
+        <hutool.version>5.5.8</hutool.version>
         <ini4j.version>0.5.4</ini4j.version>
-        <mapstruct.version>1.3.1.Final</mapstruct.version>
+        <mapstruct.version>1.4.2.Final</mapstruct.version>
         <shiro.version>1.5.1</shiro.version>
         <jwt.version>3.10.1</jwt.version>
         <guava.version>28.2-jre</guava.version>
-        <snakeyaml.version>1.26</snakeyaml.version>
+        <snakeyaml.version>1.27</snakeyaml.version>
         <swagger2.version>2.9.2</swagger2.version>
-        <knife4j.version>2.0.2</knife4j.version>
+        <knife4j.version>2.0.8</knife4j.version>
 
         <maven-compiler-plugin.version>3.8.1</maven-compiler-plugin.version>
         <maven-resources-plugin.version>3.1.0</maven-resources-plugin.version>
@@ -176,19 +180,6 @@
                 <artifactId>mysql-connector-java</artifactId>
                 <version>${mysql.version}</version>
             </dependency>
-
-            <!-- swagger start -->
-            <dependency>
-                <groupId>io.springfox</groupId>
-                <artifactId>springfox-swagger2</artifactId>
-                <version>${swagger2.version}</version>
-            </dependency>
-            <dependency>
-                <groupId>io.springfox</groupId>
-                <artifactId>springfox-swagger-ui</artifactId>
-                <version>${swagger2.version}</version>
-            </dependency>
-            <!-- swagger end -->
 
             <dependency>
                 <groupId>com.github.xiaoymin</groupId>
@@ -548,11 +539,6 @@
             <id>aliyun</id>
             <name>aliyun-maven</name>
             <url>http://maven.aliyun.com/nexus/content/groups/public/</url>
-        </repository>
-        <repository>
-            <id>spring-milestones</id>
-            <name>Spring Milestones</name>
-            <url>https://maven.aliyun.com/repository/spring</url>
         </repository>
         <repository>
             <id>central</id>
