@@ -16,14 +16,13 @@
 
 package io.geekidea.springbootplus.admin;
 
-import de.codecentric.boot.admin.server.config.EnableAdminServer;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.ConfigurableApplicationContext;
-import org.springframework.context.annotation.Configuration;
 import org.springframework.core.env.ConfigurableEnvironment;
+
+import de.codecentric.boot.admin.server.config.EnableAdminServer;
+import lombok.extern.slf4j.Slf4j;
 
 /**
  * Spring Boot Admin 启动类
@@ -32,8 +31,6 @@ import org.springframework.core.env.ConfigurableEnvironment;
  * @date 2020/3/20
  **/
 @Slf4j
-@Configuration
-@EnableAutoConfiguration
 @EnableAdminServer
 @SpringBootApplication
 public class SpringBootPlusAdminApplication {
